refactor(data-model): rename `json-encoding.ts` to `impl.ts`
`codec-json` already names its shared vocabulary `interface.ts`; `impl.ts`
is the parallel name for the area's main implementation, so the pair reads
as one convention rather than two unrelated file names. The test file
follows, and its top-level `describe()` with it.
